scripts/requirements.txt
boto3==1.36.14
numpy==2.4.2
pandas==3.0.1
pyyaml==6.0.3
requests==2.32.5
